change friend modal style to picture style
