+++ src/component/journal/AppBar.js
@@ -1,0 +1,39 @@
+import AppBar from "@material-ui/core/AppBar";
+import CheckIcon from "@material-ui/icons/Check";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import { makeStyles } from "@material-ui/core/styles";
+import IconButton from "@material-ui/core/IconButton";
+import React from "react";
+
+const useStyles = makeStyles((theme) => ({
+  menuButton: {
+    marginRight: theme.spacing(2),
+  },
+}));
+
+export default function MyAppBar({ onSubmit, string }) {
+  const classes = useStyles();
+
+  return (
+    <div>
+      <AppBar position="static">
+        <Toolbar variant="dense">
+          <Typography variant="h6" color="inherit" style={{ flexGrow: 1 }}>
+            {string}
+          </Typography>
+
+          <IconButton
+            edge="end"
+            className={classes.menuButton}
+            color="inherit"
+            aria-label="menu"
+            onClick={onSubmit}
+          >
+            <CheckIcon />
+          </IconButton>
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
